nlp_service/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# NLP & Machine Learning
torch>=1.13.0
transformers>=4.25.0
sentence-transformers>=2.2.0
sentencepiece>=0.1.97
tokenizers>=0.13.0

# API & Web
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
python-multipart>=0.0.6

# OCR & Image processing
pytesseract>=0.3.10
Pillow>=9.4.0
opencv-python>=4.7.0  # Optional, for advanced image processing

# Utilities
matplotlib>=3.5.0
seaborn>=0.12.0
tqdm>=4.64.0
psutil>=5.9.0
python-dotenv>=0.21.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.23.0
black>=22.0.0  # Optional, for code formatting
flake8>=6.0.0  # Optional, for linting

# Windows specific (if needed)
# pywin32>=305  # For Windows-specific OCR optimizations